# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Voitity API",
        "description": "API documentation for Voitity",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/avatar/generate": {
            "post": {
                "tags": [
                    "Avatar"
                ],
                "summary": "Generate an avatar from a profile image",
                "operationId": "bfd9dd469a926b8706832276d35b262d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "profile_id",
                                    "image"
                                ],
                                "properties": {
                                    "profile_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar generation started successfully."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Profile not found"
                    },
                    "402": {
                        "description": "Subscription limit exceeded"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": [
                            "avatar:write"
                        ]
                    }
                ]
            }
        },
        "/api/avatar/{profile}": {
            "get": {
                "tags": [
                    "Avatar"
                ],
                "summary": "Get active profile avatar",
                "operationId": "a5e3971c1d9a8a113c9dbcb54f489ce2",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar retrieved successfully."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Profile or avatar not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": [
                            "avatar:read"
                        ]
                    }
                ]
            }
        },
...
    },
...
}